# gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Gateway для взаимодействия с file-storing и file-analysis",
        "title": "HSE Helper Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/analyze": {
            "post": {
                "description": "Analyze the content of a file and return the analysis result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File Analysis"
                ],
                "summary": "Analyze file content",
                "parameters": [
                    {
                        "description": "Analyze request",
                        "name": "analyzeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis successful",
                        "schema": {
                            "$ref": "#/definitions/services.AnalyzeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/files": {
            "get": {
                "description": "Fetch a list of all files from file-storing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File Get"
                ],
                "summary": "List all files",
                "responses": {
                    "200": {
                        "description": "List of files",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.FileResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Fetch a file from file-storing by its ID and stream it to the client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File Get"
                ],
                "summary": "Get file by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/reports/{id}": {
            "get": {
                "description": "Fetch an analysis report from file-analysis by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report Get"
                ],
                "summary": "Get analysis report by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/services.Report"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Upload a file and analyze its content, saving the result and byte stream to file-storing",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File Upload"
                ],
                "summary": "Upload and analyze file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File upload and analysis successful",
                        "schema": {
                            "$ref": "#/definitions/services.UploadAnalysisResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "services.AnalyzeRequest": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "services.AnalyzeResponse": {
            "type": "object",
            "properties": {
                "characters": {
                    "type": "integer"
                },
                "is_plagiarized": {
                    "type": "boolean"
                },
                "paragraphs": {
                    "type": "integer"
                },
                "words": {
                    "type": "integer"
                }
            }
        },
        "services.FileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "services.Report": {
            "type": "object",
            "properties": {
                "characters": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paragraphs": {
                    "type": "integer"
                },
                "words": {
                    "type": "integer"
                }
            }
        },
        "services.UploadAnalysisResponse": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/services.AnalyzeResponse"
                },
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                }
            }
        }
    }
}